Refactor KV Offloading
