+++ vllm/v1/kv_offload/cpu/spec.py
@@ -5,13 +5,41 @@ from collections.abc import Iterator
 from vllm.config import VllmConfig
 from vllm.platforms import current_platform
 from vllm.v1.kv_cache_interface import KVCacheConfig
-from vllm.v1.kv_offload.abstract import LoadStoreSpec, OffloadingManager
-from vllm.v1.kv_offload.cpu.manager import CPUOffloadingManager
-from vllm.v1.kv_offload.mediums import CPULoadStoreSpec, GPULoadStoreSpec
-from vllm.v1.kv_offload.reuse_manager import FilterReusedOffloadingManager
-from vllm.v1.kv_offload.spec import CanonicalKVCaches, OffloadingSpec
-from vllm.v1.kv_offload.worker.cpu_gpu import CpuGpuOffloadingHandlers
-from vllm.v1.kv_offload.worker.worker import OffloadingHandler
+from vllm.v1.kv_offload.base import (
+    BlockIDsLoadStoreSpec,
+    CanonicalKVCaches,
+    GPULoadStoreSpec,
+    LoadStoreSpec,
+    OffloadingManager,
+    OffloadingSpec,
+)
+
+# -----------------------------------------------------------------------------
+# CPULoadStoreSpec — defined here to avoid circular imports with cpu/manager.py
+# (cpu/manager.py imports CPULoadStoreSpec from this module)
+# -----------------------------------------------------------------------------
+
+
+class CPULoadStoreSpec(BlockIDsLoadStoreSpec):
+    """
+    Spec for loading/storing a KV block to CPU memory.
+    """
+
+    @staticmethod
+    def medium() -> str:
+        return "CPU"
+
+
+# ----------------------------------------------------------------------------
+# CPUOffloadingSpec — imports cpu/manager.py after CPULoadStoreSpec is defined
+# ----------------------------------------------------------------------------
+
+from vllm.v1.kv_offload.cpu.gpu_worker import CpuGpuOffloadingHandlers  # noqa: E402
+from vllm.v1.kv_offload.cpu.manager import (  # noqa: E402
+    CPUOffloadingManager,
+    FilterReusedOffloadingManager,
+)
+from vllm.v1.kv_offload.worker.worker import OffloadingHandler  # noqa: E402
 
 
 class CPUOffloadingSpec(OffloadingSpec):
